bugfix in asr/kaldi/src/latgen_lib.cc

+++ asr/utils/misc.py
@@ -38,12 +38,12 @@ def int2onehot(idx, num_classes, floor=0.):
     value = 1. - floor * (num_classes - 1)
     assert value > floor
     if not torch.is_tensor(idx):
-        onehot = idx.new_full((1, num_classes), floor)
+        onehot = idx.new_full((1, num_classes), floor, dtype=torch.float)
         idx = torch.LongTensor([idx])
         onehot.scatter_(1, idx.unsqueeze(0), value)
     else:
         sizes = idx.size()
-        onehot = idx.new_full((idx.numel(), num_classes), floor)
+        onehot = idx.new_full((idx.numel(), num_classes), floor, dtype=torch.float)
         onehot.scatter_(1, idx.view(-1).long().unsqueeze(1), value)
         onehot = onehot.view(*sizes, -1)
     return onehot
